Add files via upload
The dataset is collected from UCI Machine Learning Repository in Comma Separated Value (CSV) format. It contains rows of Customers Information which are stored in a single column. The main issues is that the data is not parsed into separate columns and the values in it has some characters such as semicolons and quotation marks these make it inappropriate for analysis. Therefore, the above data cleaning processes centered on separating the dataset into multiple columns and these characters are removed in order to have an analytic dataset.